modified state.h to support samuel

+++ src/gamePlayer.cpp
@@ -11,7 +11,7 @@ GamePlayer::GamePlayer(Colour colour, int ply, int row, int column) {
 
 string GamePlayer::play() {
     string optimalMove = "";
-    int val = minimax(0, this->currentState, true, this->ply, optimalMove, INT32_MIN, INT32_MAX, this->currentState->colourOfCurrentPlayer);
+    double val = minimax(0, this->currentState, true, this->ply, optimalMove, double(INT32_MIN), double(INT32_MAX), this->currentState->colourOfCurrentPlayer);
 
     // changing board of current state
     this->currentState->makeMove(optimalMove, this->currentState->currentBoard);
